Revert "View component for sourcediff on Request#show"
