+++ src/api/app/views/webui/request/request_action.js.erb
@@ -1,1 +1,1 @@
-$('.tab-content.sourcediff').append("<%= escape_javascript(render(SourcediffTabComponent.new(bs_request: @bs_request, action: @action, active: @active, index: @index, refresh: @refresh))) %>");
+$('.tab-content.sourcediff').append("<%= escape_javascript(render(partial: 'sourcediff_tab', locals: { bs_request: @bs_request, action: @action, active: @active, index: @index, refresh: @refresh })) %>");
