+++ src/api/spec/controllers/webui/webui_controller_spec.rb
@@ -94,6 +94,29 @@ RSpec.describe Webui::WebuiController do
     end
   end
 
+  describe '#valid_xml_id' do
+    it "ensures that xml_id starts with '_' or a character" do
+      expect(controller.valid_xml_id('123')).to eq('_123')
+      expect(controller.valid_xml_id('abc')).to eq('abc')
+    end
+
+    it 'substitutes invalid characters with underscores' do
+      expect(controller.valid_xml_id('abc+&: .()~@#')).to eq('abc__________')
+    end
+
+    it 'html escapes the input' do
+      expect(controller.valid_xml_id('foo<bar&>?')).to eq('foo&lt;bar_&gt;?')
+    end
+
+    it 'leaves valid characters untouched' do
+      expect(controller.valid_xml_id('aA1-?%$§{}[]\=|')).to eq('aA1-?%$§{}[]\=|')
+    end
+
+    it 'prepends an underscore if id does not start with a valid character' do
+      expect(controller.valid_xml_id('10.2')).to eq('_10_2')
+    end
+  end
+
   describe '#set_project before filter' do
     context 'with invalid project parameter' do
       it 'raises an ActiveRecord::RecordNotFound exception' do
